requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.0
colorama==0.4.6
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.1
git-filter-repo==2.47.0
groq==0.34.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ics==0.7.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sniffio==1.3.1
TatSu==5.13.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
